requirements.txt
dl-myo
numpy
pyqtgraph
PySide6